src/main/java/com/cleverdeveloper/gof/behavioral/command/baeldung/invoker/TextFileOperationExecutor.java
package com.cleverdeveloper.gof.behavioral.command.baeldung.invoker;

/*
PROJECT NAME : Back 2 Basics (Java Core)
Module NAME: IntelliJ IDEA
Author Name : @ DRRONIDZ
DATE : 1/16/2023 11:21 PM
*/

import com.cleverdeveloper.gof.behavioral.command.baeldung.command.TextFileOperation;

import java.util.ArrayList;
import java.util.List;

public class TextFileOperationExecutor {

    private final List<TextFileOperation> textFileOperations = new ArrayList<>();

    public String executeOperation(TextFileOperation textFileOperation) {
        textFileOperations.add(textFileOperation);
        return textFileOperation.execute();
    }
}
